# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin-ms/v1/merchanat/audit-logs": {
            "get": {
                "description": "Retrieve audit logs with optional filtering by user, service, and time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuditLog"
                ],
                "summary": "Get Audit Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "service",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/data.AuditLog"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin-ms/v1/merchant/audit-logs/verify": {
            "get": {
                "description": "Verify audit logs within a specified time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuditLog"
                ],
                "summary": "Verify Audit Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.AuditLogVerifyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin-ms/v1/merchant/risk-user-reviews": {
            "get": {
                "description": "Retrieve a paginated list of risk user reviews with optional filtering by status, user ID and creation time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RiskUserReview"
                ],
                "summary": "List Risk User Reviews",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by decision value (tinyint)",
                        "name": "decision",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Filter create_time \u003e= start_time (unix timestamp)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Filter create_time \u003c= end_time (unix timestamp)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.RiskUserReviewListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin-ms/v1/merchant/risk-user-reviews/{review_id}/decision": {
            "put": {
                "description": "Update the decision and optional decision_source for a risk user review identified by user_id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RiskUserReview"
                ],
                "summary": "Update Risk User Review Decision",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "record ID of the risk user review to update",
                        "name": "review_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update decision request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateDecisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.AuditLog": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "description": "ID of the user who performed the action",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "data.AuditLogVerifyResponse": {
            "type": "object",
            "properties": {
                "failed_log_id": {
                    "type": "string"
                },
                "is_valid": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "data.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "err_msg": {
                    "type": "string"
                }
            }
        },
        "data.RiskUserReview": {
            "type": "object",
            "properties": {
                "analyst_summary": {
                    "type": "string"
                },
                "confidence": {
                    "type": "string"
                },
                "create_time": {
                    "type": "integer"
                },
                "decision": {
                    "type": "integer"
                },
                "decision_source": {
                    "type": "string"
                },
                "fraud_probability": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "risk_level": {
                    "type": "string"
                },
                "risk_score": {
                    "type": "number"
                },
                "rule_score": {
                    "type": "number"
                },
                "rules": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "data.RiskUserReviewListResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.RiskUserReview"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "data.UpdateDecisionRequest": {
            "type": "object",
            "required": [
                "id",
                "user_id"
            ],
            "properties": {
                "decision": {
                    "type": "integer"
                },
                "decision_source": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}